pyproject.toml
[project]
name = "oursappback"
version = "0.1.0"
description = "Ours app backend."
authors = [
    {name = "beichenstanly",email = "zhengxgen@163.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
